Inline NoMethodsInPropertySetup
