--- Source/Moq.Analyzers.Test/Data/NoMethodsInPropertySetup.Bad.cs
@@ -1,24 +1,0 @@
-using Moq;
-
-namespace NoMethodsInPropertySetup.Bad;
-
-public interface IFoo
-{
-    string Prop1 { get; set; }
-
-    string Prop2 { get; }
-
-    string Prop3 { set; }
-
-    string Method();
-}
-
-public class MyUnitTests
-{
-    private void TestBad()
-    {
-        var mock = new Mock<IFoo>();
-        mock.SetupGet(x => x.Method());
-        mock.SetupSet(x => x.Method());
-    }
-}
